Save Gig to DB

+++ Models/Genre.cs
@@ -8,7 +8,7 @@ namespace GigHub.Models
 {
     public class Genre
     {
-        public int Id { get; set; }
+        public byte Id { get; set; }
         [Required,StringLength(255)]
         public string Name { get; set; }
     }
